weighted yield
Failed attempt. I thought weighting the yield from regime at each prospective S by how likely we were to observe that S under each regime would help account for the fac that we will likely never be in regime 2 w high S. This failed bc there is not yield to maximize under regime 2 so it should not be a target. This is demonstrated in lines 1157-1192 of SET.R