Polish quarto
Thought I was done with quarto and ready to share but still troubled by poor estimation of gamma (spans the prior in most cases unless fishing at 50% of MSY at the lb). Also think the gamma prior is overly influential. That might be ok but should probably be more explicit about it. Should do more to identify when the SET becomes reliable.