minor edits
Want to start a new branch where we drop lnalpha_red to be a percentage of the lb since to document when we can estimate well. Focus on pctMSY at the lb isn’t answering the question cleanly. Also, pctMSY at the lb is better for describing relationship between SET and lb but I’m not sure we need to carry that over to estimation.