Modified sim_SRgamma for pinks. Also, checked/corrected indexing. Updated script and quarto with age-at-maturity scenarios.
At this time sim_SRgamma still has SOC designation internally although it is unused and refers to an early SOC policy recommendation (3 levels but management and conservation based on run sizes). Should fix that next (1 level based on run size). Questionable if we need to do this inside the function. Are we convinced recovery goals don’t make sense? Do we need to demonstrate that?